HBP-113 Moved data_encryption_handler.py under core.entities.file; Fixed project renaming
